Add day 10

+++ util/template.cpp
@@ -10,14 +10,14 @@ auto parse( const std::string& input )
     return getNumbers< int >( input, ',' );
 }
 
-int partOne( const std::string& input )
+unsigned long long partOne( const std::string& input )
 {
     parse( input );
 
     return 0;
 }
 
-int partTwo( const std::string& input )
+unsigned long long partTwo( const std::string& input )
 {
     parse( input );
